Cleanup schema definitions
