+++ ui/src/api/namespaces/mutate/createNamespace.ts
@@ -1,4 +1,4 @@
-import { MirrorPostSchemaType } from "~/api/namespaces/schema/mirror";
+import { MirrorPostPatchSchemaType } from "~/api/namespaces/schema/mirror";
 import { NamespaceCreatedEditedSchema } from "../schema/namespace";
 import type { NamespaceListSchemaType } from "../schema/namespace";
 import { apiFactory } from "~/api/apiFactory";
@@ -32,7 +32,7 @@ export const useCreateNamespace = ({
       mirror,
     }: {
       name: string;
-      mirror?: MirrorPostSchemaType;
+      mirror?: MirrorPostPatchSchemaType;
     }) =>
       createNamespace({
         apiKey: apiKey ?? undefined,
